rt_block_paging
